requirements.txt
streamlit 
nltk 
sentence-transformers 
scikit-learn 
textblob 
pyyaml
groq 
pandas 
altair